requirements.txt
streamlit==1.56.0
pandas==3.0.2
mitreattack-python==5.4.4
openai==2.31.0
google-generativeai==0.8.6
openpyxl==3.1.5
requests==2.33.1
stix2==3.0.2
fpdf2==2.8.7
SQLAlchemy==2.0.49
PyYAML==6.0.3
pytest==9.0.3
plotly==6.6.0
PyPDF2==3.0.1
beautifulsoup4==4.14.3
fastapi==0.135.3
uvicorn[standard]==0.44.0

